+++ zarrs_metadata/src/v3/metadata.rs
@@ -54,6 +54,12 @@ pub struct MetadataV3 {
     must_understand: bool,
 }
 
+impl Into<Option<Configuration>> for MetadataV3 {
+    fn into(self) -> Option<Configuration> {
+        self.configuration
+    }
+}
+
 impl TryFrom<&str> for MetadataV3 {
     type Error = serde_json::Error;