+++ zarrs/src/lib.rs
@@ -42,7 +42,9 @@
 //!   - Experimental extensions are either pending registration or have no formal specification outside of the `zarrs` docs.
 //!   - Experimental extensions may be unrecognised or incompatible with other Zarr implementations.
 //!   - Experimental extensions may change in future releases without maintaining backwards compatibility.
-//! - *Deprecated*: indicated by ~~strikethrough~~ in the tables below, and are not recommended for use in new arrays.
+//! - *Deprecated*: indicated by ~~strikethrough~~ in the tables below
+//!   - Deprecated aliases will not be removed, but are not recommended for use in new arrays.
+//!   - Deprecated extensions may be removed in future releases.
 //!
 //! Extension names and aliases are configurable with [`Config::codec_aliases_v3_mut`](config::Config::codec_aliases_v3_mut) and similar methods for data types and Zarr V2.
 //! `zarrs` will persist extension names if opening an existing array of creating an array from metadata.